fastapi_agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "AG-UI compliant agent service for chat"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.114.2",
    "openai<2.0.0,>=1.76.0",
    "ag-ui-protocol>=0.1.5",
    "pydantic>2.0",
    "langchain-openai==0.2.3",
    "langchain-community>=0.3.1",
    "langgraph-checkpoint-postgres==2.0.21",
    "langgraph==0.5.4",
    "loguru==0.7.3",
    "tavily-python==0.7.1"
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "pytest-asyncio<1.0.0,>=0.23.0",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
